Moved Dataview to its own component
