+++ zappnode-front/src/components/dashboard/BarChart.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function BarChart() {
+  return <div>barchart</div>;
+}
+
+export default BarChart;
